# Updated to firstSliceSize=4, maxFirstSliceSize=64