What's the best feature of deque? IMO it's its design. It is just so flexible that allows us to tweak deque in all sorts of ways.

Below tweaks boost the performance very nicely for small data sets while still managing to keep or improving the performance for larger data sets.

```
benchstat testdata/BenchmarkMicroserviceDequeQueueMaster.txt testdata/BenchmarkMicroserviceDequeQueuePR.txt
name        old time/op    new time/op    delta
/1-4           503ns ± 1%     457ns ± 0%   -9.06%  (p=0.000 n=8+10)
/10-4         3.50µs ± 4%    2.77µs ± 1%  -20.84%  (p=0.000 n=9+10)
/100-4        23.5µs ± 1%    24.2µs ± 1%   +2.97%  (p=0.000 n=10+10)
/1000-4        230µs ± 4%     224µs ± 1%   -2.70%  (p=0.000 n=10+10)
/10000-4      2.30ms ± 3%    2.30ms ± 4%     ~     (p=0.447 n=9+10)
/100000-4     24.9ms ± 2%    24.6ms ± 0%   -1.46%  (p=0.000 n=9+9)
/1000000-4     260ms ± 1%     258ms ± 1%   -0.98%  (p=0.019 n=10+10)

name        old alloc/op   new alloc/op   delta
/1-4            560B ± 0%      544B ± 0%   -2.86%  (p=0.000 n=10+10)
/10-4         5.71kB ± 0%    2.58kB ± 0%  -54.90%  (p=0.000 n=10+10)
/100-4        15.8kB ± 0%    20.9kB ± 0%  +32.62%  (p=0.000 n=10+10)
/1000-4        133kB ± 0%     134kB ± 0%   +0.76%  (p=0.000 n=10+10)
/10000-4      1.43MB ± 0%    1.43MB ± 0%   +0.07%  (p=0.000 n=10+10)
/100000-4     14.4MB ± 0%    14.4MB ± 0%   +0.01%  (p=0.000 n=9+10)
/1000000-4     144MB ± 0%     144MB ± 0%   +0.00%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
/1-4            12.0 ± 0%      11.0 ± 0%   -8.33%  (p=0.000 n=10+10)
/10-4           77.0 ± 0%      75.0 ± 0%   -2.60%  (p=0.000 n=10+10)
/100-4           707 ± 0%       709 ± 0%   +0.28%  (p=0.000 n=10+10)
/1000-4        7.01k ± 0%     7.01k ± 0%     ~     (all equal)
/10000-4       70.2k ± 0%     70.2k ± 0%     ~     (all equal)
/100000-4       702k ± 0%      702k ± 0%     ~     (all equal)
/1000000-4     7.02M ± 0%     7.02M ± 0%     ~     (all equal)

benchstat testdata/BenchmarkMicroserviceDequeStackMaster.txt testdata/BenchmarkMicroserviceDequeStackPR.txt
name        old time/op    new time/op    delta
/1-4           399ns ± 4%     349ns ± 1%  -12.62%  (p=0.000 n=10+9)
/10-4         2.65µs ± 5%    2.48µs ± 0%   -6.57%  (p=0.000 n=10+10)
/100-4        22.9µs ± 1%    23.0µs ± 0%   +0.36%  (p=0.010 n=9+10)
/1000-4        220µs ± 1%     219µs ± 1%     ~     (p=0.089 n=10+10)
/10000-4      2.25ms ± 0%    2.23ms ± 0%   -0.86%  (p=0.000 n=10+8)
/100000-4     25.2ms ± 5%    24.5ms ± 0%   -2.96%  (p=0.000 n=10+9)
/1000000-4     271ms ±12%     257ms ± 1%   -5.17%  (p=0.006 n=10+9)

name        old alloc/op   new alloc/op   delta
/1-4            304B ± 0%      288B ± 0%   -5.26%  (p=0.000 n=10+10)
/10-4         1.57kB ± 0%    1.55kB ± 0%   -1.02%  (p=0.000 n=10+10)
/100-4        15.8kB ± 0%    16.8kB ± 0%   +6.38%  (p=0.000 n=10+10)
/1000-4        129kB ± 0%     130kB ± 0%   +0.78%  (p=0.000 n=10+10)
/10000-4      1.43MB ± 0%    1.42MB ± 0%   -0.51%  (p=0.000 n=10+10)
/100000-4     14.4MB ± 0%    14.4MB ± 0%   +0.01%  (p=0.000 n=8+9)
/1000000-4     144MB ± 0%     144MB ± 0%   +0.00%  (p=0.000 n=9+10)

name        old allocs/op  new allocs/op  delta
/1-4            11.0 ± 0%      10.0 ± 0%   -9.09%  (p=0.000 n=10+10)
/10-4           75.0 ± 0%      74.0 ± 0%   -1.33%  (p=0.000 n=10+10)
/100-4           707 ± 0%       707 ± 0%     ~     (all equal)
/1000-4        7.01k ± 0%     7.01k ± 0%     ~     (all equal)
/10000-4       70.2k ± 0%     70.2k ± 0%   -0.01%  (p=0.000 n=10+10)
/100000-4       702k ± 0%      702k ± 0%     ~     (all equal)
/1000000-4     7.02M ± 0%     7.02M ± 0%     ~     (all equal)
```

There's tons of possible combinations between tweaking firstSliceSize, sliceGrowthFactor, maxFirstSliceSize, maxInternalSliceSize and maxSpareLinks.

The million dollar question is: is there a better combination that could yield better performance? (very likely, yes!)

For now the process to find out the best combination is a manual one. We're working on an automation that will automatically run and find out the best combinations.